+++ MatchEntities/Models/WeaponReload.cs
@@ -11,11 +11,11 @@ namespace MatchEntities
         public int Time { get; set; }
         public int Tick { get; set; }
         public long PlayerId { get; set; }
-        public double PlayerPosX { get; set; }
-        public double PlayerPosY { get; set; }
-        public double PlayerPosZ { get; set; }
-        public double PlayerViewX { get; set; }
-        public double PlayerViewY { get; set; }
+        public float PlayerPosX { get; set; }
+        public float PlayerPosY { get; set; }
+        public float PlayerPosZ { get; set; }
+        public float PlayerViewX { get; set; }
+        public float PlayerViewY { get; set; }
         public bool IsCt { get; set; }
         public short Weapon { get; set; }
         public short AmmoBefore { get; set; }
